Add toggle for custom marks on waterfall chart and update design per the feedback.
