+++ x-pack/plugins/synthetics/public/apps/synthetics/components/step_details_page/step_waterfall_chart/waterfall/waterfall_chart_wrapper.tsx
@@ -6,6 +6,7 @@
  */
 
 import React, { useCallback, useMemo, useState } from 'react';
+import useLocalStorage from 'react-use/lib/useLocalStorage';
 import { EuiHealth } from '@elastic/eui';
 import { JourneyStep, NetworkEvent } from '../../../../../../../common/runtime_types';
 import { getSeriesAndDomain, getSidebarItems } from '../../common/network_data/data_formatting';
@@ -40,7 +41,14 @@ export const WaterfallChartWrapper: React.FC<Props> = ({
 }) => {
   const [query, setQuery] = useState<string>('');
   const [activeFilters, setActiveFilters] = useState<string[]>([]);
-  const [onlyHighlighted, setOnlyHighlighted] = useState(false);
+  const [onlyHighlighted, setOnlyHighlighted] = useLocalStorage<boolean>(
+    'xpack.synthetics.waterfallChart.showOnlyHighlighted',
+    false
+  );
+  const [showCustomMarks, setShowCustomMarks] = useLocalStorage<boolean>(
+    'xpack.synthetics.waterfallChart.showCustomMarks',
+    false
+  );
 
   const [networkData] = useState<NetworkEvent[]>(data);
 
@@ -51,7 +59,7 @@ export const WaterfallChartWrapper: React.FC<Props> = ({
   }, [networkData, query, activeFilters, onlyHighlighted]);
 
   const sidebarItems = useMemo(() => {
-    return getSidebarItems(networkData, onlyHighlighted, query, activeFilters);
+    return getSidebarItems(networkData, onlyHighlighted ?? false, query, activeFilters);
   }, [networkData, query, activeFilters, onlyHighlighted]);
 
   const {
@@ -100,12 +108,14 @@ export const WaterfallChartWrapper: React.FC<Props> = ({
       onElementClick={useCallback(onBarClick, [onBarClick])}
       onProjectionClick={useCallback(onProjectionClick, [onProjectionClick])}
       onSidebarClick={onSidebarClick}
-      showOnlyHighlightedNetworkRequests={onlyHighlighted}
+      showOnlyHighlightedNetworkRequests={onlyHighlighted ?? false}
+      showCustomMarks={showCustomMarks ?? false}
       sidebarItems={sidebarItems}
       metadata={metadata}
       activeFilters={activeFilters}
       setActiveFilters={setActiveFilters}
       setOnlyHighlighted={setOnlyHighlighted}
+      setShowCustomMarks={setShowCustomMarks}
       query={query}
       setQuery={setQuery}
       renderTooltipItem={useCallback((tooltipProps) => {
