+++ x-pack/plugins/synthetics/public/apps/synthetics/components/step_details_page/step_waterfall_chart/waterfall/waterfall_marker/waterfall_markers.tsx
@@ -23,7 +23,7 @@ export const LAYOUT_SHIFT = 'layoutShift';
 
 export function WaterfallChartMarkers() {
   const theme = useTheme();
-  const { markerItems } = useWaterfallContext();
+  const { markerItems, showCustomMarks } = useWaterfallContext();
 
   const markerItemsByOffset = useMemo(
     () =>
@@ -38,8 +38,10 @@ export function WaterfallChartMarkers() {
     let lastOffset: number;
     const recognizedMarkerItemsByOffset = Array.from(markerItemsByOffset.entries()).reduce(
       (acc, [offset, items]) => {
-        // Remove unrecognized marks e.g. custom marks
-        const vitalMarkers = items.filter(({ id }) => getMarkersInfo(id, theme) !== undefined);
+        // Remove unrecognized marks e.g. custom marks if `showCustomMarks` is false
+        const vitalMarkers = showCustomMarks
+          ? items
+          : items.filter(({ id }) => getMarkersInfo(id, theme) !== undefined);
 
         const hasMultipleMarksAtOffset = vitalMarkers.some(({ id }) => id !== LAYOUT_SHIFT);
         const isLastOffsetTooClose = lastOffset && Math.abs(offset - lastOffset) < 100; // 100ms
@@ -79,7 +81,7 @@ export function WaterfallChartMarkers() {
         dash: markersInfo?.dash,
       };
     });
-  }, [markerItemsByOffset, theme]);
+  }, [markerItemsByOffset, showCustomMarks, theme]);
 
   if (!markerItems) {
     return null;
