+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -35855,10 +35855,8 @@
     "xpack.synthetics.wait": "等待",
     "xpack.synthetics.waterfall.applyFilters.label": "选择要应用筛选的项目",
     "xpack.synthetics.waterfall.applyFilters.message": "单击以添加或移除筛选",
-    "xpack.synthetics.waterfall.chartLegend.heading": "图例",
     "xpack.synthetics.waterfall.clearFilters.label": "清除筛选",
     "xpack.synthetics.waterfall.networkRequests.heading": "网络请求",
-    "xpack.synthetics.waterfall.networkRequests.hideNonMatching": "隐藏不匹配项",
     "xpack.synthetics.waterfallChart.sidebar.url.https": "https",
     "xpack.threatIntelligence.common.emptyPage.body3": "要开始使用 Elastic 威胁情报，请从“集成”页面启用一个或多个威胁情报集成，或使用 Filebeat 采集数据。有关更多信息，请查看 {docsLink}。",
     "xpack.threatIntelligence.addToBlockList": "添加阻止列表条目",
